Refactored data-table folder.
