--- packages/ketchup/src/components/ketchup-data-table/ketchup-data-table.spec.ts
@@ -1,7 +1,0 @@
-import { KetchupDataTable } from './ketchup-data-table';
-
-describe('ketchup-data-table', () => {
-    it('builds', () => {
-        expect(new KetchupDataTable()).toBeTruthy();
-    });
-});
