+++ packages/ketchup/src/components/kup-data-table/kup-data-table-declarations.ts
