+++ packages/ketchup/src/components/kup-data-table/kup-data-table.spec.ts
@@ -1,0 +1,7 @@
+import { KupDataTable } from './kup-data-table';
+
+describe('kup-data-table', () => {
+    it('builds', () => {
+        expect(new KupDataTable()).toBeTruthy();
+    });
+});
