+++ packages/ketchup/src/components/kup-data-table/kup-data-table.tsx
@@ -20,23 +20,23 @@ import {
     Cell,
     RowAction,
     ShowGrid,
-} from './ketchup-data-table-declarations';
+} from './kup-data-table-declarations';
 
 import {
     calcTotals,
     filterRows,
     groupRows,
     sortRows,
-} from './ketchup-data-table-helper';
+} from './kup-data-table-helper';
 
 import { isIcon, isImage, isLink } from '../../utils/object-utils';
 
 @Component({
     tag: 'kup-data-table',
-    styleUrl: 'ketchup-data-table.scss',
+    styleUrl: 'kup-data-table.scss',
     shadow: true,
 })
-export class KetchupDataTable {
+export class KupDataTable {
     @Prop() data: { columns?: Array<Column>; rows?: Array<Row> };
 
     @Prop()
